Implement a more efficient way to check for primes

+++ src/main.rs
@@ -207,7 +207,7 @@ fn main() {
 			println!("{}", result);
 		} else if args[2] == "prime" {
 			for value in data_integer {
-				if num_integer::is_prime(value as i64) {
+				if num_integer::is_prime(value as u32) {
 					print!("True ");
 				} else {
 					print!("False ");
